diagnostic variable updates
don't use typecasts
use 16bit unsigned variables

+++ src/autosteer.cpp
@@ -107,7 +107,7 @@ void autosteerWorker100Hz( void* z ) {
           String str;
           str.reserve( 30 );
           str = "\nNumber of faults: ";
-          str += ( int8_t ) diagnostics.steerEnabledWithNoPower;
+          str += diagnostics.steerEnabledWithNoPower;
           str += "\nFault active since startup: Yes";
           diagnosticsDisplay.steerEngagedFaults = str;
           saveDiagnostics();
